Crate rome_js_formatter

Rome’s official JavaScript formatter.

Implement the formatter

Our formatter is node based. Meaning that each AST node knows how to format itself. In order to implement the formatting, a node has to implement the traitFormatNode`.

rome has an automatic code generation that creates automatically the files out of the grammar. By default, all implementations will format verbatim, meaning that the formatter will print tokens and trivia as they are (format_verbatim).

Our formatter has its own internal IR, it creates its own abstraction from an AST.

The developer won’t be creating directly this IR, but they will use a series of utilities that will help to create this IR. The whole IR is represented by the enum FormatElement. Please refer to its internal documentation to understand the meaning of each variant.

Rules to follow when implementing a formatter

Use the *Fields struct to extract all the tokens/nodes

#[derive(Debug, Clone, Default)]
pub struct FormatJsExportDefaultExpressionClause;

impl FormatNodeRule<JsExportDefaultExpressionClause> for FormatJsExportDefaultExpressionClauses {
    fn fmt_fields(&self, node: &JsExportDefaultExpressionClause, f: &mut JsFormatter) -> FormatResult<()> {
        let JsExportDefaultExpressionClauseFields {
            default_token,
            expression,
            semicolon_token,
        }  = node.as_fields();
   }
}

When using .as_fields() with the destructuring, don’t use the .. feature. Prefer extracting all fields and ignore them using the _

#[derive(Debug, Clone, Default)]
pub struct FormatJsExportDefaultExpressionClause;

impl FormatNodeRule<JsExportDefaultExpressionClause> for FormatJsExportDefaultExpressionClauses {
    fn fmt_fields(&self, node: &JsExportDefaultExpressionClause, f: &mut JsFormatter) -> FormatResult<()> {
         let JsExportDefaultExpressionClauseFields {
             default_token,
             expression: _,
             semicolon_token
         } = node.as_fields();
     }
}

The reason why we want to promote this pattern is because we want to make explicit when a token/node is excluded;

Use the APIs provided by builders.rs, formatter and format_extensions.rs.

builders.rs exposes a series of utilities to craft the formatter IR; please refer to their internal documentation to understand what the utilities are for;

formatter exposes a set of functions to help to format some recurring patterns; please refer to their internal documentation to understand how to use them and when;

format_extensions.rs: with these traits, we give the ability to nodes and tokens to implements certain methods that are exposed based on its type. If you have a good IDE support, this feature will help you. For example:

#[derive(Debug, Clone, Default)]
pub struct FormatJsExportDefaultExpressionClause;

impl FormatNodeRule<JsExportDefaultExpressionClause> for FormatJsExportDefaultExpressionClauses{
     fn fmt_fields(&self, node: &JsExportDefaultExpressionClause, f: &mut JsFormatter) -> FormatResult<()> {
         let JsExportDefaultExpressionClauseFields {
             default_token,
             expression, // it's a mandatory node
             semicolon_token, // this is not a mandatory node
         } = node.as_fields();
         let element = expression.format();

         if let Some(expression) = &expression? {
             write!(f, [expression.format(), space_token()])?;
         }

         if let Some(semicolon) = &semicolon_token {
             write!(f, [semicolon.format()])?;
         } else {
             write!(f, [space_token()])?;
         }
     }
}

Use our playground to inspect the code that you want to format. You can inspect the AST given by a certain snippet. This will help you to understand which nodes need to be implemented/modified in order to implement formatting. Alternatively, you can locally run the playground by following the playground instructions.

Use the quick_test() function to test you snippet straight from your IDE, without running the whole test suite. The test is ignored on purpose, so you won’t need to worry about the CI breaking.

Write tests for the formatter

We use insta.rs for our snapshot tests, please make sure you read its documentation to learn the basics of snapshot testing. You should install the companion cargo-insta command to assist with snapshot reviewing.

Directories are divided by language, so when creating a new test file, make sure to have the correct file under the correct folder:

JavaScript => js/ directory

TypeScript => ts/ directory

JSX => jsx/ directory

TSX => ts/ directory

To create a new snapshot test for JavaScript, create a new file to crates/rome_js_formatter/tests/specs/js/, e.g. arrow_with_spaces.js

const foo     = ()    => {
    return bar
}

Files processed as modules must go inside the module/ directory, files processed as script must go inside the script/ directory.

Run the following command to generate the new snapshot (the snapshot tests are generated by a procedure macro so we need to recompile the tests):

touch crates/rome_js_formatter/tests/spec_tests.rs && cargo test -p rome_js_formatter formatter

For better test driven development flow, start the formatter tests with cargo-watch:

cargo watch -i '*.new' -x 'test -p rome_js_formatter formatter'

After test execution, you will get a new arrow.js.snap.new file.

To actually update the snapshot, run cargo insta review to interactively review and accept the pending snapshot. arrow.js.snap.new will be replaced with arrow.js.snap

Sometimes, you need to verify the formatting for different cases/options. In order to do that, create a folder with the cases you need to verify. If we needed to follow the previous example:

create a folder called arrow_with_spaces/ and move the JS file there;

then create a file called options.json

The content would be something like:

{
    "cases": [
        {
            "line_width": 120,
            "indent_style": {"Space": 4}
        }
    ]
}

the cases keyword is mandatory;

then each object of the array will contain the matrix of options you’d want to test. In this case the test suite will run a second test case with line_width to 120 and ident_style with 4 spaces

when the test suite is run, you will have two outputs in your snapshot: the default one and the custom one

Identify issues

There are four cases when a test is not correct:

you try to print/format the same token multiple times; the formatter will check at runtime when a test is run;

some tokens haven’t been printed; usually you will have this information inside the snapshot, under a section called "Unimplemented tokens/nodes"; a test, in order to be valid, can’t have that section;

If removing a token is the actual behaviour (removing some parenthesis or a semicolon), then the correct way to do it by using the formatter API formatter.format_replaced(token, empty_element());

the emitted code is not a valid program anymore, the test suite will parse again the emitted code and it will fail if there are syntax errors;

the emitted code, when formatted again, differs from the original; this usually happens when removing/adding new elements, and the grouping is not correctly set;

Write tests for a parser

If you want to create a new test for an existing parser, you will have to inline the code that you want to test in a comment that is created in a specific way.

Let’s say that you created a new parsing feature and you need new tests from scratch, just go to the source code where you parse this new feature if JavaScript, and add the following comment:

// test feature_name
// let a = { new_feature : "" }
// let b = { new_feature : "" }
fn parse_new_feature(p: &mut Parser) -> ParsedSyntax {}

The first line, // test feature_name the important one. This will tell to the testing infrastructure to create a positive test (without parsing errors), called feature_name.js inside the test_data/inline/ok folder.

The content of this file will be:

let a = { new_feature : "" }
let b = { new_feature : "" }

Basically, everything after the key comment will be the content of the new file.

Now you need to run cargo codegen test and the task will actually generate this file for you.

In case you want to create a negative test (with parsing errors), you will create a new comment like this:

// test feature_name
// let a = { new_feature : "" }
// let b = { new_feature : "" }

+ // test_err feature_name
+ // let a = {  : "" }
+ // let b = { new_feature :  }
fn parse_new_feature(p: &mut Parser) -> ParsedSyntax {}

Mind the different comment test_err, which marks the error for the test suite as a test that has to fail.

Run the command cargo codegen test and you will see a new file called feature_name.js inside the test_data/inline/err folder.

The content of this file will be:

let a = {  : "" }
let b = { new_feature :  }

Now run the command: Unix/macOS

env UPDATE_EXPECT=1 cargo test

Windows

set UPDATE_EXPECT=1 & cargo test

The command will tell the test suite to generate and update the .rast files.

If tests that are inside the ok/ folder fail or if tests that are inside the err/ folder don’t emit, the whole test suite will fail.

Modules

comments

context

prelude

This module provides important and useful traits to help to format tokens and nodes when implementing the crate::FormatNodeRule trait.

utils

Structs

FormatJsSyntaxToken

Format implementation specific to JavaScript tokens.

FormattedIter

Traits

AsFormat

Used to get an object that knows how to format this object.

FormatNodeRule

Rule for formatting a JavaScript AstNode.

FormatUnknownNodeRule

Rule for formatting an unknown node.

FormattedIterExt

Formatting specific Iterator extensions

IntoFormat

Used to convert this object into an object that can be formatted.

Functions

format_node

Formats a JavaScript (and its super languages) file based on its features.

format_range

Formats a range within a file, supported by Rome

format_sub_tree

Formats a single node within a file, supported by Rome.