Struct rome_js_formatter::context::JsFormatContext

pub struct JsFormatContext { /* private fields */ }

Implementations

impl JsFormatContext

pub fn new(source_type: SourceType) -> Self

pub fn with_indent_style(self, indent_style: IndentStyle) -> Self

pub fn with_line_width(self, line_width: LineWidth) -> Self

pub fn with_quote_style(self, quote_style: QuoteStyle) -> Self

pub fn with_source_type(self, source_type: SourceType) -> Self

pub fn line_width(&self) -> LineWidth

pub fn quote_style(&self) -> QuoteStyle

pub fn source_type(&self) -> SourceType

pub fn is_suppressed(&mut self, node: &JsSyntaxNode) -> bool

Returns true if the passed node has a suppression comment and tracks in debug builds that the suppression comments of this node have been checked.

Examples

use rome_js_formatter::context::JsFormatContext;
use rome_js_formatter::prelude::*;
use rome_js_parser::parse_expression;
use rome_js_syntax::SourceType;

let root = parse_expression(r#"
// rome-ignore format: Suppressed for testing purposes
console.log('abcd');
"#, 0).tree();

let call_expression = root.expression().unwrap();
let mut context = JsFormatContext::new(SourceType::js_module());

assert!(context.is_suppressed(call_expression.syntax()));

impl JsFormatContext

pub fn tab_width(&self) -> TabWidth