Function rome_js_formatter::prelude::format_once

pub const fn format_once<T, Context>(formatter: T) -> FormatOnce<T, Context> where
    T: FnOnce(&mut Formatter<'_, Context>) -> Result<(), FormatError>, 

Creates an inline Format object that can only be formatted once.

This can be useful in situation where the borrow checker doesn’t allow you to use format_with because the code formatting the content consumes the value and cloning the value is too expensive. An example of this is if you want to nest a FormatElement or non-cloneable Iterator inside of a block_indent as shown can see in the examples section.

Panics

Panics if the object gets formatted more than once.

Example

use rome_formatter::prelude::*;
use rome_formatter::{SimpleFormatContext, format, write, Buffer};

struct MyFormat;

fn generate_values() -> impl Iterator<Item=StaticText> {
    vec![text("1"), text("2"), text("3"), text("4")].into_iter()
}

impl Format<SimpleFormatContext> for MyFormat {
    fn fmt(&self, f: &mut Formatter<SimpleFormatContext>) -> FormatResult<()> {
        let mut values = generate_values();

        let first = values.next();

        // Formats the first item outside of the block and all other items inside of the block,
        // separated by line breaks
        write!(f, [
            first,
            block_indent(&format_once(|f| {
                // Using format_with isn't possible here because the iterator gets consumed here
                f.join_with(&hard_line_break()).entries(values).finish()
            })),
        ])
    }
}

let formatted = format!(SimpleFormatContext::default(), [MyFormat])?;

assert_eq!("1\n\t2\n\t3\n\t4\n", formatted.print()?.as_code());

Formatting the same value twice results in a panic.

use rome_formatter::prelude::*;
use rome_formatter::{SimpleFormatContext, format, write, Buffer};
use rome_rowan::TextSize;

let mut count = 0;

let value = format_once(|f| {
    write!(f, [dynamic_token(&std::format!("Formatted {count}."), TextSize::default())])
});

format!(SimpleFormatContext::default(), [value]).expect("Formatting once works fine");

// Formatting the value more than once panics
format!(SimpleFormatContext::default(), [value]);