Function rome_js_formatter::prelude::format_or_verbatim

pub const fn format_or_verbatim<'a, Node>(
    node: &'a Node
) -> FormatNodeOrVerbatim<'a, Node> where
    Node: AstNode<Language = JsLanguage> + AsFormat<'a>, 

Formats a node using its AsFormat implementation but falls back to printing the node as it is in the source document if the formatting returns an FormatError.