Function rome_js_formatter::prelude::soft_space_or_block_indent

pub fn soft_space_or_block_indent<Context>(
    content: &impl Format<Context>
) -> BlockIndent<'_, Context>

Adds spaces around the content if its enclosing group fits on a line, otherwise indents the content and separates it by line breaks.

Examples

Adds line breaks and indents the content if the enclosing group doesn’t fit on the line.

use rome_formatter::{format, format_args, LineWidth, SimpleFormatOptions};
use rome_formatter::prelude::*;

let context = SimpleFormatContext::new(SimpleFormatOptions {
    line_width: LineWidth::try_from(10).unwrap(),
    ..SimpleFormatOptions::default()
});

let elements = format!(context, [
    group(&format_args![
        text("{"),
        soft_space_or_block_indent(&format_args![
            text("aPropertyThatExceeds"),
            text(":"),
            space(),
            text("'line width'"),
        ]),
        text("}")
    ])
])?;

assert_eq!(
    "{\n\taPropertyThatExceeds: 'line width'\n}",
    elements.print()?.as_code()
);

Adds spaces around the content if the group fits on the line

use rome_formatter::{format, format_args};
use rome_formatter::prelude::*;

let elements = format!(SimpleFormatContext::default(), [
    group(&format_args![
        text("{"),
        soft_space_or_block_indent(&format_args![
            text("a"),
            text(":"),
            space(),
            text("5"),
        ]),
        text("}")
    ])
])?;

assert_eq!(
    "{ a: 5 }",
    elements.print()?.as_code()
);