Function rome_js_formatter::prelude::syntax_token_cow_slice

pub fn syntax_token_cow_slice<L>(
    text: Cow<'a, str>,
    token: &'a SyntaxToken<L>,
    start: TextSize
) -> SyntaxTokenCowSlice<'a, L> where
    L: Language, 

String that is the same as in the input source text if text is Cow::Borrowed or some replaced content if text is Cow::Owned.