Function rome_js_formatter::prelude::token

pub fn token(text: &'static str) -> StaticToken

Creates a token that gets written as is to the output. Make sure to properly escape the text if it’s user generated (e.g. a string and not a language keyword).

Line feeds

Tokens may contain line breaks but they must use the line feeds (\n). The crate::Printer converts the line feed characters to the character specified in the crate::PrinterOptions.

Examples

use rome_formatter::format;
use rome_formatter::prelude::*;

let elements = format!(SimpleFormatContext::default(), [token("Hello World")]).unwrap();

assert_eq!(
    "Hello World",
    elements.print().as_code()
);

Printing a string literal as a literal requires that the string literal is properly escaped and enclosed in quotes (depending on the target language).

use rome_formatter::format;
use rome_formatter::prelude::*;

// the tab must be encoded as \\t to not literally print a tab character ("Hello{tab}World" vs "Hello\tWorld")
let elements = format!(SimpleFormatContext::default(), [token("\"Hello\\tWorld\"")]).unwrap();

assert_eq!(r#""Hello\tWorld""#, elements.print().as_code());