Struct rome_js_formatter::prelude::IfGroupBreaks

pub struct IfGroupBreaks<'a, Context> { /* private fields */ }

Implementations

impl<Context> IfGroupBreaks<'_, Context>

pub fn with_group_id(
self,
group_id: Option<DebugGroupId>
) -> IfGroupBreaks<'_, Context>

Inserts some content that the printer only prints if the group with the specified group_id is printed in multiline mode. The referred group must appear before this element in the document but doesn’t have to one of its ancestors.

Examples

Prints the trailing comma if the array group doesn’t fit. The group_id is necessary because fill creates an implicit group around each item and tries to print the item in flat mode. The item [4] in this example fits on a single line but the trailing comma should still be printed

use rome_formatter::{format, format_args, write, LineWidth, SimpleFormatOptions};
use rome_formatter::prelude::*;

let context = SimpleFormatContext::new(SimpleFormatOptions {
    line_width: LineWidth::try_from(20).unwrap(),
    ..SimpleFormatOptions::default()
});

let formatted = format!(context, [format_with(|f| {
    let group_id = f.group_id("array");

    write!(f, [
        group(
            &format_args![
                text("["),
                soft_block_indent(&format_with(|f| {
                    f.fill()
                        .entry(&soft_line_break_or_space(), &text("1,"))
                        .entry(&soft_line_break_or_space(), &text("234568789,"))
                        .entry(&soft_line_break_or_space(), &text("3456789,"))
                        .entry(&soft_line_break_or_space(), &format_args!(
                            text("["),
                            soft_block_indent(&text("4")),
                            text("]"),
                            if_group_breaks(&text(",")).with_group_id(Some(group_id))
                        ))
                    .finish()
                })),
                text("]")
            ],
        ).with_group_id(Some(group_id))
    ])
})])?;

assert_eq!(
    "[\n\t1, 234568789,\n\t3456789, [4],\n]",
    formatted.print()?.as_code()
);