Enum rome_js_formatter::prelude::tag::Tag

pub enum Tag {

    StartIndent,
    EndIndent,
    StartAlign(Align),
    EndAlign,
    StartDedent(DedentMode),
    EndDedent,
    StartGroup(Group),
    EndGroup,
    StartConditionalContent(Condition),
    EndConditionalContent,
    StartIndentIfGroupBreaks(DebugGroupId),
    EndIndentIfGroupBreaks,
    StartFill,
    EndFill,
    StartEntry,
    EndEntry,
    StartLineSuffix,
    EndLineSuffix,
    StartVerbatim(VerbatimKind),
    EndVerbatim,
    StartLabelled(LabelId),
    EndLabelled,

}

A Tag marking the start and end of some content to which some special formatting should be applied.

Tags always come in pairs of a start and an end tag and the styling defined by this tag will be applied to all elements in between the start/end tags.

Variants

StartIndent

Indents the content one level deeper, see crate::builders::indent for documentation and examples.

EndIndent

StartAlign(Align)

Variant of TagKind::Indent that indents content by a number of spaces. For example, Align(2) indents any content following a line break by an additional two spaces.

Nesting (Aligns)TagKind::Align has the effect that all except the most inner align are handled as (Indent)TagKind::Indent.

EndAlign

StartDedent(DedentMode)

Reduces the indention of the specified content either by one level or to the root, depending on the mode. Reverse operation of Indent and can be used to undo an Align for nested content.

EndDedent

StartGroup(Group)

Creates a logical group where its content is either consistently printed:

on a single line: Omitting LineMode::Soft line breaks and printing spaces for LineMode::SoftOrSpace

on multiple lines: Printing all line breaks

See crate::builders::group for documentation and examples.

EndGroup

StartConditionalContent(Condition)

Allows to specify content that gets printed depending on whatever the enclosing group is printed on a single line or multiple lines. See crate::builders::if_group_breaks for examples.

EndConditionalContent

StartIndentIfGroupBreaks(DebugGroupId)

Optimized version of Tag::StartConditionalContent for the case where some content should be indented if the specified group breaks.

EndIndentIfGroupBreaks

StartFill

Concatenates multiple elements together with a given separator printed in either flat or expanded mode to fill the print width. Expect that the content is a list of alternating [element, separator] See crate::Formatter::fill.

EndFill

StartEntry

Entry inside of a Tag::StartFill

EndEntry

StartLineSuffix

Delay the printing of its content until the next line break

EndLineSuffix

StartVerbatim(VerbatimKind)

A token that tracks tokens/nodes that are printed as verbatim.

EndVerbatim

StartLabelled(LabelId)

Special semantic element marking the content with a label. This does not directly influence how the content will be printed.

See crate::builders::labelled for documentation.

EndLabelled

Implementations

impl Tag

pub const fn is_start(&self) -> bool

Returns true if self is any start tag.

pub const fn is_end(&self) -> bool

Returns true if self is any end tag.

pub const fn kind(&self) -> TagKind