Trait rome_js_formatter::prelude::FormatAstSeparatedListExtension

pub trait FormatAstSeparatedListExtension: AstSeparatedList<Language = JsLanguage> {
    fn format_separated(
        &self,
        separator: &'static str
    ) -> FormatSeparatedIter<AstSeparatedListElementsIterator<JsLanguage, Self::Node>, JsLanguage, Self::Node> { ... }
}

AST Separated list formatting extension methods

Provided Methods

fn format_separated(
&self,
separator: &'static str
) -> FormatSeparatedIter<AstSeparatedListElementsIterator<JsLanguage, Self::Node>, JsLanguage, Self::Node>

Prints a separated list of nodes

Trailing separators will be reused from the original list or created by calling the separator_factory function. The last trailing separator in the list will only be printed if the outer group breaks.