Trait rome_js_formatter::prelude::MemoizeFormat

pub trait MemoizeFormat<Context> {
    fn memoized(self) -> Memoized<Self, Context>
    where
        Self: Format<Context>,
    { ... }
}

Utility trait that allows memorizing the output of a Format. Useful to avoid re-formatting the same object twice.

Provided Methods

fn memoized(self) -> Memoized<Self, Context>where
Self: Format<Context>,

Returns a formattable object that memoizes the result of Format by cloning. Mainly useful if the same sub-tree can appear twice in the formatted output because it’s used inside of if_group_breaks or if_group_fits_single_line.

use std::cell::Cell;
use rome_formatter::{format, write};
use rome_formatter::prelude::*;
use rome_rowan::TextSize;

struct MyFormat {
  value: Cell<u64>
}

impl MyFormat {
    pub fn new() -> Self {
        Self { value: Cell::new(1) }
    }
}

impl Format<SimpleFormatContext> for MyFormat {
    fn fmt(&self, f: &mut Formatter<SimpleFormatContext>) -> FormatResult<()> {
        let value = self.value.get();
        self.value.set(value + 1);

        write!(f, [dynamic_text(&std::format!("Formatted {value} times."), TextSize::from(0))])
    }
}

let normal = MyFormat::new();

// Calls `format` for everytime the object gets formatted
assert_eq!(
    "Formatted 1 times. Formatted 2 times.",
    format!(SimpleFormatContext::default(), [normal, space(), normal])?.print()?.as_code()
);

// Memoized memoizes the result and calls `format` only once.
let memoized = normal.memoized();
assert_eq!(
    "Formatted 3 times. Formatted 3 times.",
    format![SimpleFormatContext::default(), [memoized, space(), memoized]]?.print()?.as_code()
);