Trait rome_js_formatter::prelude::_

pub trait _<Context> {
    type Target: Format<Context>;

    fn with_or_empty<With>(
        self,
        with: With
    ) -> Option<FormatItemWith<With, Self::Target>>
    where
        With: Fn(&Self::Target, &mut Formatter<'_, Context>) -> Result<(), FormatError>;
}

Utility trait used to simplify the formatting of optional objects that are formattable.

Required Associated Types

type Target: Format<Context>

Required Methods

fn with_or_empty<With>(
self,
with: With
) -> Option<FormatItemWith<With, Self::Target>>where
With: Fn(&Self::Target, &mut Formatter<'_, Context>) -> Result<(), FormatError>,

This function tries to format an optional object. If the object is None nothing is written in the buffer. If exists, the utility formats the object and passes it to the closure.

Examples

use rome_formatter::{write, format};
use rome_formatter::prelude::*;
use rome_rowan::TextSize;

struct MyFormat;

impl Format<SimpleFormatContext> for MyFormat {
    fn fmt(&self, f: &mut Formatter<SimpleFormatContext>) -> FormatResult<()> {
        write!(f, [text("MyToken")])
    }
}

let none_token: Option<MyFormat> = None;
// Returns `empty_element()` for a `None` value
let none_formatted = format!(SimpleFormatContext::default(), [
    none_token.with_or_empty(|token, f| write!(f, [token]))
]).unwrap();

assert!(none_formatted.into_document().is_empty());

let some_token = Some(MyFormat);
assert_eq!(
    format![SimpleFormatContext::default(), [space(), text("MyToken")]],
    format!(
        SimpleFormatContext::default(), [
            some_token.with_or_empty(|token, f| {
                write!(f, [space(), token])
            })
        ]
    )
);