Trait rome_js_formatter::AsFormat

pub trait AsFormat {
    type Format<'a>: Format<JsFormatContext>
    where
        Self: 'a;

    fn format(&self) -> Self::Format<'_>;
}

Used to get an object that knows how to format this object.

Required Associated Types

type Format<'a>: Format<JsFormatContext>
where
Self: 'a

Required Methods

fn format(&self) -> Self::Format<'_>

Returns an object that is able to format this object.