Trait rome_js_formatter::FormatNodeRule

pub trait FormatNodeRule<N> where
    N: AstNode<Language = JsLanguage>, {
    fn fmt_fields(
        &self,
        item: &N,
        f: &mut Formatter<'_, JsFormatContext>
    ) -> FormatResult<()>;

    fn fmt(
        &self,
        node: &N,
        f: &mut Formatter<'_, JsFormatContext>
    ) -> FormatResult<()> { ... }
    fn fmt_node(
        &self,
        node: &N,
        f: &mut Formatter<'_, JsFormatContext>
    ) -> FormatResult<()> { ... }
    fn needs_parentheses(&self, item: &N) -> bool { ... }
    fn is_suppressed(&self, node: &N, f: &Formatter<'_, JsFormatContext>) -> bool { ... }
    fn fmt_leading_comments(
        &self,
        node: &N,
        f: &mut Formatter<'_, JsFormatContext>
    ) -> FormatResult<()> { ... }
    fn fmt_dangling_comments(
        &self,
        node: &N,
        f: &mut Formatter<'_, JsFormatContext>
    ) -> FormatResult<()> { ... }
    fn fmt_trailing_comments(
        &self,
        node: &N,
        f: &mut Formatter<'_, JsFormatContext>
    ) -> FormatResult<()> { ... }
}

Rule for formatting a JavaScript AstNode.

Required Methods

fn fmt_fields(
&self,
item: &N,
f: &mut Formatter<'_, JsFormatContext>
) -> FormatResult<()>

Formats the node’s fields.

Provided Methods

fn fmt(
&self,
node: &N,
f: &mut Formatter<'_, JsFormatContext>
) -> FormatResult<()>

fn fmt_node(
&self,
node: &N,
f: &mut Formatter<'_, JsFormatContext>
) -> FormatResult<()>

Formats the node without comments. Ignores any suppression comments.

fn needs_parentheses(&self, item: &N) -> bool

Returns whether the node requires parens.

fn is_suppressed(&self, node: &N, f: &Formatter<'_, JsFormatContext>) -> bool

Returns true if the node has a suppression comment and should use the same formatting as in the source document.

fn fmt_leading_comments(
&self,
node: &N,
f: &mut Formatter<'_, JsFormatContext>
) -> FormatResult<()>

Formats the leading comments of the node.

You may want to override this method if you want to manually handle the formatting of comments inside of the fmt_fields method or customize the formatting of the leading comments.

fn fmt_dangling_comments(
&self,
node: &N,
f: &mut Formatter<'_, JsFormatContext>
) -> FormatResult<()>

Formats the dangling comments of the node.

You should override this method if the node handled by this rule can have dangling comments because the default implementation formats the dangling comments at the end of the node, which isn’t ideal but ensures that no comments are dropped.

A node can have dangling comments if all its children are tokens or if all node childrens are optional.

fn fmt_trailing_comments(
&self,
node: &N,
f: &mut Formatter<'_, JsFormatContext>
) -> FormatResult<()>

Formats the trailing comments of the node.

You may want to override this method if you want to manually handle the formatting of comments inside of the fmt_fields method or customize the formatting of the trailing comments.