Module rome_js_formatter::utils::jsx_utils

Enums

WrapState

Indicates that an element should always be wrapped in parentheses, should be wrapped only when it’s line broken, or should not be wrapped at all.

Statics

JSX_WHITESPACE_CHARS

Functions

contains_meaningful_jsx_text

Checks if the children of an element contain meaningful text. See is_meaningful_jsx_text for definition of meaningful JSX text.

get_wrap_state

Checks if a JSX Element should be wrapped in parentheses. Returns a WrapState which indicates when the element should be wrapped in parentheses.

is_jsx_inside_arrow_function_inside_call_inside_expression_child

This is a very special situation where we’re returning a JsxElement from an arrow function that’s passed as an argument to a function, which is itself inside a JSX expression child.

is_meaningful_jsx_text

Meaningful JSX text is defined to be text that has either non-whitespace characters, or does not contain a newline. Whitespace is defined as ASCII whitespace.