Trait rome_js_formatter::AsFormat

pub trait AsFormat<'a> {
    type Format: Format<JsFormatContext>;

    fn format(&'a self) -> Self::Format;
}

Used to get an object that knows how to format this object.

Required Associated Types

type Format: Format<JsFormatContext>

Required Methods

fn format(&'a self) -> Self::Format

Returns an object that is able to format this object.