Trait rome_js_formatter::IntoFormat

pub trait IntoFormat<Context> {
    type Format: Format<Context>;

    fn into_format(self) -> Self::Format;
}

Used to convert this object into an object that can be formatted.

The difference to AsFormat is that this trait takes ownership of self.

Required Associated Types

type Format: Format<Context>

Required Methods

fn into_format(self) -> Self::Format